% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareApportionmentApproaches.R
\docType{data}
\name{angle_total_deg}
\alias{angle_total_deg}
\title{This script does some comparisons between different apportionment approaches.}
\format{An object of class \code{numeric} of length 1.}
\usage{
angle_total_deg
}
\description{
This script does some comparisons between different apportionment approaches.
}
\keyword{datasets}
