% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion_web.R
\name{apportion_web}
\alias{apportion_web}
\title{Distribute streamflow depletion within a stream network using web distance weighting.}
\usage{
apportion_web(reach_dist, w, max_dist)
}
\arguments{
\item{reach_dist}{data frame with two columns: \code{reach}, which is a grouping variable with
the name of each stream reach, and \code{dist} which is the distance of a point on that stream reach to
the well of interest. There can (and likely will) be more than one \code{dist} per \code{reach};
if there is only one dist per reach, results will be the same as the \link{apportion_inverse} method.}

\item{w}{weighting factor; 1 for inverse distance, 2 for inverse distance squared.}

\item{max_dist}{the maximum distance of a stream to be depleted}
}
\value{
A data frame with two columns:
\describe{
  \item{reach}{the grouping variable input in \code{reach_dist}}
  \item{frac_depletion}{the proportion of streamflow depletion from the well occurring in that reach.}
}
}
\description{
Since analytical models assume the presence of 1 (or sometimes 2) linear streams,
functions are needed to distribute that depletion to various reaches within a
real stream network. These geometric functions are described in Zipper et al (2018).
}
\examples{
# these will be the same as apportion_inverse because there's only one dist per reach
reach_dist <- data.frame(reach = seq(1,5),
  dist = c(100, 150, 900, 300, 200))
apportion_web(reach_dist, w = 2, max_dist = 1000)
apportion_web(reach_dist, w = 2, max_dist = 500)

reach_dist <- data.frame(reach = c("A", "A", "A", "B", "B"),
  dist = c(100, 150, 900, 300, 200))
apportion_web(reach_dist, w = 1, max_dist = 1000)
apportion_web(reach_dist, w = 1, max_dist = 500)
}
\references{
Zipper, SC, T Dallemagne, T Gleeson, TC Boerman, A Hartmann (2018). Groundwater Pumping Impacts
on Real Stream Networks: Testing the Performance of Simple Management Tools. Water Resources Research.
doi:10.1029/2018WR022707.
}
