% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion_wedge.R
\name{apportion_wedge}
\alias{apportion_wedge}
\title{Distribute streamflow depletion between two streams in a wedge-shaped aquifer.}
\usage{
apportion_wedge(angle_total, angle_well)
}
\arguments{
\item{angle_total}{angle [radians] between the two streams.}

\item{angle_well}{angle [radians] from the first (lower boundary) stream and the well.}
}
\value{
A numeric of length two with the proportion of steady-state
capture fraction from the first (lower) and second (upper) streams.
}
\description{
Streams are two linear tributaries that meet at the origin. From Yeh et al. (2008), Eq. 18 & 19.
}
\examples{
apportion_wedge(angle_total=(50*pi/2), angle_well=(10*pi/2))

apportion_wedge(angle_total=(50*pi/2), angle_well=(25*pi/2))
}
\references{
Yeh, HD, YC Chang, VA Zlotnik (2008). Stream depletion rate and volume from groundwater pumping in
wedge-shaped aquifers. Journal of Hydrology 349(3): 501-511. doi:10.1029/2018WR022707.
}
