% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush.R
\name{hantush}
\alias{hantush}
\title{Hantush (1965) analytical model for streamflow depletion with partially penetrating stream with semipervious streambed.}
\usage{
hantush(t, d, S, Kh, b, Kriv, briv, prec = 80)
}
\arguments{
\item{t}{times you want output for [T]}

\item{d}{distance from well to stream [L]}

\item{S}{aquifer storage coefficient (specific yield if unconfined; storativity if confined)}

\item{Kh}{aquifer horizontal hydraulic conductivity [L/T]}

\item{b}{aquifer saturated thickness [L]}

\item{Kriv}{streambed semipervious layer hydraulic conductivity [L/T]}

\item{briv}{streambed semipervious layer thickness [L]}

\item{prec}{precision for mpfr package for storing huge numbers; 80 seems to generally work but tweak this if you get weird results.

Reference:
Hantush, MS (1965). Wells near Streams with Semipervious Beds. Journal of Geophysical Research 70(12): 2829-38. doi:10.1029/JZ070i012p02829.
 
As the leakance term (b*Kh/Kriv) approaches 0 (thin streambed, high Kriv) this is equivalent to glover method

Output:
  Qf = streamflow depletion as fraction of pumping rate [-]
  
If you have the pumping rate of the well [Qw; L3/T] you can
calculate the rate of streamflow depletion [Qs; L3/T] as Qs=Qf*Qw
  
Assumptions (from Reeves et al., 2009):
 -Horizontal flow >> vertical flow (Dupuit assumptions hold)
 -Homogeneous, isotropic aquifer
 -Aquifer is confined, or if unconfined change in head is small relative to total thickness (constant Tr)
 -Stream is straight, infinitely long, and remains in hydraulic connection to aquifer
 -Pumping does not change the stage of the stream  
 -Recharge to the system is unchanged by pumping
 -Streambed may offer resistance to groundwater flow
 -No streambank storage
 -Pumping rate is constant
 -Aquifer extends to infinity
 
Example run code is at bottom of this .R file (below function)}
}
\description{
Hantush (1965) analytical model for streamflow depletion with partially penetrating stream with semipervious streambed.
}
