% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittentPumping.R
\name{intermittentPumping}
\alias{intermittentPumping}
\title{intermittentPumping}
\usage{
intermittentPumping(t, starts, stops, rates, method = "glover", d, S, Tr, ...)
}
\arguments{
\item{t}{times you want output for [T]}

\item{starts}{vector of times to start pumping [T] (must be same length as stops and rates)}

\item{stops}{vector of times pumping stops [T] (must be same length as starts and rates)}

\item{rates}{vector of pumping rates [L3/T] (must be same length as starts and stops)}

\item{method}{analytical solution to use (options= 'glover', 'hunt'). Defaults to 'glover'.}

\item{d}{distance from well to stream [L]}

\item{S}{aquifer storage coefficient (specific yield if unconfined; storativity if confined)}

\item{Tr}{aquifer transmissivity [L2/T]}

\item{...}{any other inputs required for your \code{method} of choice; for example, \code{hunt} needs \code{lmda} (streambed conductance)}
}
\description{
Calculate streamflow depletion for an intermittent pumping schedule using the Jenkins (1968) superposition technique.
}
\examples{
intermittentPumping(t=seq(0,60,10), starts=0, stops=30, rates=100, method="glover", d=100, S=0.1, Tr=100)

Reference: 
Jenkins, C.T. (1968). Techniques for Computing Rate and Volume of Stream Depletion
by Wells. Ground Water 6(2): 37-46. doi:10.1111/j.1745-6584.1968.tb01641.x

Output:
  Q = streamflow depletion [L3/T]
         Note that this is NOT Qf (depletion fraction) because depletion fraction
         would not work if there were different pumping rates at different times
}
