% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_lines}
\alias{stream_lines}
\title{Stream network for Sixmile Creek Watershed, Wisconsin, USA. Extracted from US NHDPlus v2.1 national seamless dataset.}
\format{
A simple feature LINESTRONG collection with 49 rows and 2 variables:
\describe{
  \item{reach}{identifier code for each stream reach}
  \item{stream}{name of stream for each stream reach (Sixmile Creek or Dorn Creek)}
  ...
}
}
\source{
\url{http://www.horizon-systems.com/NHDPlusData/NHDPlusV21/Data/NationalData/}
}
\usage{
stream_lines
}
\description{
Stream network for Sixmile Creek Watershed, Wisconsin, USA. Extracted from US NHDPlus v2.1 national seamless dataset.
}
\keyword{datasets}
