% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_sf}
\alias{stream_sf}
\title{Stream network for Sixmile Creek Watershed, WI.}
\format{A simple LINESTRING feature collection with 49 features and 2 fields:
\describe{
  \item{reach}{identifier code for each stream reach}
  \item{stream}{name of stream for each stream reach (Sixmile Creek or Dorn Creek)}
  ...
}}
\source{
\url{http://www.horizon-systems.com/NHDPlusData/NHDPlusV21/Data/NationalData/NHDPlusV21_NationalData_CONUS_Seamless_Geodatabase_05.7z}
}
\usage{
stream_sf
}
\description{
Stream network for Sixmile Creek Watershed, WI.
}
\keyword{datasets}
