% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{subtract.bounded}
\alias{subtract.bounded}
\title{Contains little functions that may be called from other functions.}
\usage{
subtract.bounded(x, y, lower.bound = -Inf, upper.bound = Inf)
}
\description{
Contains little functions that may be called from other functions.
}
