% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{subtract_bounded}
\alias{subtract_bounded}
\title{Contains little functions that may be called from other functions.
subtraction bounded with lower/upper limits}
\usage{
subtract_bounded(x, y, lower_bound = -Inf, upper_bound = Inf)
}
\arguments{
\item{x}{first term in subtraction}

\item{y}{second term in subtraction}

\item{lower_bound}{minimum allowed result. Default to Inf.}

\item{upper_bound}{maximum allowed result. Default to -Inf.}
}
\value{
(x-y), within the limits of lower_bound and upper_bound
}
\description{
Called from \code{intermittentPumping}
}
\examples{
subtract_bounded(10, 15)  # returns -5; default bounds are \\code{-Inf, Inf}
subtract_bounded(10, 15, lower_bound = 0)  # returns 0 due to lower_bound
}
