% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glover.R
\name{glover}
\alias{glover}
\title{Glover and Balmer (1954) analytical model for streamflow depletion with fully-penetrating stream.}
\usage{
glover(t, d, S, Tr)
}
\arguments{
\item{t}{times you want output for [T]}

\item{d}{distance from well to stream [L]}

\item{S}{aquifer storage coefficient (specific yield if unconfined; storativity if confined)}

\item{Tr}{aquifer transmissivity [L2/T]}
}
\value{
A numeric or vector containing streamflow depletion as a fraction of pumping rate.
}
\description{
Glover and Balmer (1954) analytical model for streamflow depletion with fully-penetrating stream.
}
\examples{
glover(t=157770000, d=1000,  S=0.2, Tr=0.1)  # Glover & Balmer (1954) Table 1, Well 1
glover(t=157770000, d=5000,  S=0.2, Tr=0.1)  # Glover & Balmer (1954) Table 1, Well 2
glover(t=157770000, d=10000,  S=0.2, Tr=0.1)  # Glover & Balmer (1954) Table 1, Well 3


Reference:
Glover, RE, and GG Balmer (1954).River Depletion Resulting from Pumping a Well near a River. 
Eos, Transactions American Geophysical Union 35(3): 468-70. doi:10.1029/TR035i003p00468.
 
Output:
  Qf = streamflow depletion as fraction of pumping rate [-]
  
If you have the pumping rate of the well [Qw; L3/T] you can
calculate the rate of streamflow depletion [Qs; L3/T] as Qs=Qf*Qw

Assumptions (from Hunt, 1999):
 -Horizontal flow >> vertical flow (Dupuit assumptions hold)
 -Homogeneous, isotropic aquifer
 -Aquifer is confined, or if unconfined change in head is small relative to total thickness (constant Tr)
 -Stream is straight, infinitely long, and remains in hydraulic connection to aquifer
 -Pumping does not change the stage of the stream  
 -Recharge to the system is unchanged by pumping
 -No streambank storage
 -Pumping rate is constant
 -Aquifer extends to infinity
 
Example run code is at bottom of this .R file (below function)
}
