% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunt.R
\name{hunt}
\alias{hunt}
\title{Streamflow depletion in partially penetrating stream with semipervious streambed.}
\usage{
hunt(t, d, S, Tr, lmda, prec = 80)
}
\arguments{
\item{t}{times you want output for [T]}

\item{d}{distance from well to stream [L]}

\item{S}{aquifer storage coefficient (specific yield if unconfined; storativity if confined)}

\item{Tr}{aquifer transmissivity [L2/T]}

\item{lmda}{streambed conductance term, lambda [L/T]. Can be estimated with \code{streambed_conductance}.}

\item{prec}{precision for \code{Rmpfr} package for storing huge numbers; 80 seems to generally work but tweak this if you get weird results. Reducing this value will reduce accuracy but speed up computation time.}
}
\value{
\code{Qf}, numeric or vector of streamflow depletion as fraction of pumping rate [-]. 
If the pumping rate of the well (\code{Qw}; [L3/T]) is known, you can calculate volumetric streamflow depletion [L3/T] as \code{Qf*Qw}
}
\description{
Described in Hunt (1999). When \code{lmda} term gets very large, this is equivalent to \code{glover}.
Assumptions:
\itemize{
  \item Horizontal flow >> vertical flow (Dupuit assumptions hold)
  \item Homogeneous, isotropic aquifer
  \item Constant \code{Tr}: Aquifer is confined, or if unconfined change in head is small relative to aquifer thickness
  \item Stream is straight, infinitely long, and remains in hydraulic connection to aquifer
  \item Constant stream stage
  \item No changes in recharge due to pumping
  \item No streambank storage
  \item Constant pumping rate
  \item Aquifer extends to infinity
}
}
\references{
Hunt, B (1999). Unsteady Stream Depletion from Ground Water Pumping. 
Ground Water 37 (1): 98-102. doi:10.1111/j.1745-6584.1999.tb00962.x.
}
