% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunt.R
\name{hunt}
\alias{hunt}
\title{Hunt (1999) analytical model for streamflow depletion with a partially-penetrating stream.}
\usage{
hunt(t, d, S, Tr, lmda, prec = 80)
}
\arguments{
\item{t}{times you want output for [T]}

\item{d}{distance from well to stream [L]}

\item{S}{aquifer storage coefficient (specific yield if unconfined; storativity if confined)}

\item{Tr}{aquifer transmissivity [L2/T]}

\item{lmda}{streambed conductance term, lambda [L/T]}

\item{prec}{precision for mpfr package for storing huge numbers; 80 seems to generally work but tweak this if you get weird results. Reducing this value will reduce accuracy but speed up computation time.

Reference:
Hunt, B (1999). Unsteady Stream Depletion from Ground Water Pumping. 
Ground Water 37 (1): 98-102. doi:10.1111/j.1745-6584.1999.tb00962.x.

As the lmda term gets very high, this is equivalent to the glover method.

See script: Hunt1999_lmda to estimate lmda based on aquifer and well properties
 
Output:
  Qf = streamflow depletion as fraction of pumping rate [-]
  
If you have the pumping rate of the well [Qw; L3/T] you can
calculate the rate of streamflow depletion [Qs; L3/T] as Qs=Qf*Qw
  
Assumptions (from Reeves et al., 2009):
 -Horizontal flow >> vertical flow (Dupuit assumptions hold)
 -Homogeneous, isotropic aquifer
 -Aquifer is confined, or if unconfined change in head is small relative to total thickness (constant Tr)
 -Stream is straight, infinitely long, and remains in hydraulic connection to aquifer
 -Pumping does not change the stage of the stream  
 -Recharge to the system is unchanged by pumping
 -Streambed may offer resistance to groundwater flow
 -No streambank storage
 -Pumping rate is constant
 -Aquifer extends to infinity
 
Example run code is at bottom of this .R file (below function)}
}
\description{
Hunt (1999) analytical model for streamflow depletion with a partially-penetrating stream.
}
